COMP: Refactoring pbd collision constraint classes
1. Seperate files for separate Pbd colliison constrain classes
2. Preincrement lop variables
3. Refactor collision constraint classes: remove raw pointers, using auto, style corrections
4. Move pbd constraints to seperate folder within constraint folder
5. Clears warnings in Physics module

+++ Base/Constraint/PbdConstraints/imstkPbdConstraint.h
@@ -17,7 +17,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 
-   =========================================================================*/
+=========================================================================*/
 
 #ifndef IMSTK_PBD_CONSTRAINT_H
 #define IMSTK_PBD_CONSTRAINT_H
@@ -67,7 +67,7 @@ public:
     /// \param model \class PositionBasedDynamicsModel
     /// \return true if succeeded
     ///
-    virtual bool updateConstraint(PositionBasedDynamicsModel &model)
+    virtual bool updateConstraint(PositionBasedDynamicsModel& model)
     {
         return true;
     }
@@ -77,7 +77,7 @@ public:
     /// \param model \class PositionBasedDynamicsModel
     /// \return true if succeeded
     ///
-    virtual bool solvePositionConstraint(PositionBasedDynamicsModel &model)
+    virtual bool solvePositionConstraint(PositionBasedDynamicsModel& model)
     {
         return true;
     }
@@ -87,13 +87,13 @@ public:
     /// \param model \class PositionBasedDynamicsModel
     /// \return true if succeeded
     ///
-    virtual bool solveVelocityConstraint(PositionBasedDynamicsModel &model)
+    virtual bool solveVelocityConstraint(PositionBasedDynamicsModel& model)
     {
         return true;
     }
 
 public:
-    std::vector<unsigned int> m_vertexIds; // index of points for the constraint
+    std::vector<size_t> m_vertexIds; // index of points for the constraint
 };
 
 }
