+++ Base/Geometry/Map/imstkGeometryMap.h
@@ -119,10 +119,7 @@ public:
     /// \param idx
     /// \return index of Master corresponding to the idx of Slave
     ///
-    virtual unsigned int getMapIdx(const unsigned int& idx)
-    {
-        return 0;
-    }
+    virtual size_t getMapIdx(const size_t& idx) { return 0; }
 protected:
 
     ///
